# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Qema Al Nathair Travel & Tourism API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /package-bookings:
    get:
      operationId: listMyPackageBookings
      tags: [packages]
      summary: List current user's package bookings
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PackageBooking"
    post:
      operationId: createPackageBooking
      tags: [packages]
      summary: Book a tour package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageBookingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageBooking"
  /package-bookings/{id}:
    get:
      operationId: getPackageBooking
      tags: [packages]
      summary: Get a package booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageBooking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/package-bookings:
    get:
      operationId: listAllPackageBookings
      tags: [admin]
      summary: List all package bookings (admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/PackageBookingStatus"
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PackageBooking"
  /admin/package-bookings/{id}/status:
    patch:
      operationId: updatePackageBookingStatus
      tags: [admin]
      summary: Update package booking status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageBookingStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageBooking"

  # ---------- Flights ----------
  /flights/search:
    get:
      operationId: searchFlights
      tags: [flights]
      summary: Search flight offers
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
        - name: to
          in: query
          required: true
          schema:
            type: string
        - name: departDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: returnDate
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: tripType
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TripType"
        - name: adults
          in: query
          required: false
          schema:
            type: integer
        - name: children
          in: query
          required: false
          schema:
            type: integer
        - name: infants
          in: query
          required: false
          schema:
            type: integer
        - name: cabinClass
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/CabinClass"
      responses:
        "200":
          description: Flight offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlightOffer"
  /flight-bookings:
    get:
      operationId: listMyFlightBookings
      tags: [flights]
      summary: List current user's flight bookings
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlightBooking"
    post:
      operationId: createFlightBooking
      tags: [flights]
      summary: Book a flight offer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlightBookingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlightBooking"
  /flight-bookings/{id}:
    get:
      operationId: getFlightBooking
      tags: [flights]
      summary: Get a flight booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlightBooking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/flight-bookings:
    get:
      operationId: listAllFlightBookings
      tags: [admin]
      summary: List all flight bookings (admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/FlightBookingStatus"
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlightBooking"
  /admin/flight-bookings/{id}/status:
    patch:
      operationId: updateFlightBookingStatus
      tags: [admin]
      summary: Update flight booking status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlightBookingStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlightBooking"

  # ---------- Notifications ----------
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /admin/notifications:
    get:
      operationId: listAllNotifications
      tags: [admin]
      summary: List all sent notifications (admin)
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
    post:
      operationId: sendNotification
      tags: [admin]
      summary: Send a notification (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # ---------- Banners ----------
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ItineraryDay:
      type: object
      required: [day, title, description]
      properties:
        day: { type: integer }
        title: { type: string }
        description: { type: string }
    Package:
      type: object
      required:
        [
          id,
          name,
          country,
          city,
          days,
          nights,
          priceFrom,
          currency,
          rating,
          images,
          description,
          hotelsIncluded,
          hotelStars,
          roomType,
          meals,
          transportation,
          itinerary,
          includedServices,
          excludedServices,
          cancellationPolicy,
          isFeatured,
          createdAt,
        ]
      properties:
        id: { type: integer }
        name: { type: string }
        country: { type: string }
        city: { type: string }
        days: { type: integer }
        nights: { type: integer }
        priceFrom: { type: number }
        currency: { type: string }
        rating: { type: number }
        images:
          type: array
          items: { type: string }
        videoUrl: { type: ["string", "null"] }
        description: { type: string }
        hotelsIncluded:
          type: array
          items: { type: string }
        hotelStars: { type: integer }
        roomType: { type: string }
        meals: { type: string }
        transportation: { type: string }
        itinerary:
          type: array
          items:
            $ref: "#/components/schemas/ItineraryDay"
        includedServices:
          type: array
          items: { type: string }
        excludedServices:
          type: array
          items: { type: string }
        cancellationPolicy: { type: string }
        isFeatured: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    PackageBookingStatus:
      type: string
      enum:
        [
          received,
          reviewing,
          confirmed,
          awaiting_payment,
          paid,
          vouchers_issued,
          completed,
          cancelled,
        ]
    PackageBooking:
      type: object
      required:
        [
          id,
          userId,
          packageId,
          referenceNumber,
          travelersCount,
          travelerNames,
          passportNumbers,
          phone,
          email,
          travelDate,
          status,
          createdAt,
        ]
      properties:
        id: { type: integer }
        userId: { type: integer }
        packageId: { type: integer }
        package:
          oneOf:
            - $ref: "#/components/schemas/Package"
            - type: "null"
        referenceNumber: { type: string }
        travelersCount: { type: integer }
        travelerNames:
          type: array
          items: { type: string }
        passportNumbers:
          type: array
          items: { type: string }
        phone: { type: string }
        email: { type: string }
        travelDate: { type: string, format: date }
        notes: { type: ["string", "null"] }
        status:
          $ref: "#/components/schemas/PackageBookingStatus"
        createdAt: { type: string, format: date-time }
    PackageBookingInput:
      type: object
      required:
        [
          packageId,
          travelersCount,
          travelerNames,
          passportNumbers,
          phone,
          email,
          travelDate,
        ]
      properties:
        packageId: { type: integer }
        travelersCount: { type: integer, minimum: 1 }
        travelerNames:
          type: array
          items: { type: string }
        passportNumbers:
          type: array
          items: { type: string }
        phone: { type: string }
        email: { type: string }
        travelDate: { type: string, format: date }
        notes: { type: string }
    PackageBookingStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/PackageBookingStatus"
    TripType:
      type: string
      enum: [one_way, round_trip, multi_city]
    CabinClass:
      type: string
      enum: [economy, premium_economy, business, first]
    FlightOffer:
      type: object
      required:
        [
          id,
          airlineName,
          airlineLogoUrl,
          flightNumber,
          fromAirport,
          toAirport,
          departTime,
          arriveTime,
          durationMinutes,
          stops,
          cabinClass,
          price,
          currency,
        ]
      properties:
        id: { type: string }
        airlineName: { type: string }
        airlineLogoUrl: { type: string }
        flightNumber: { type: string }
        fromAirport: { type: string }
        toAirport: { type: string }
        departTime: { type: string, format: date-time }
        arriveTime: { type: string, format: date-time }
        durationMinutes: { type: integer }
        stops: { type: integer }
        cabinClass:
          $ref: "#/components/schemas/CabinClass"
        price: { type: number }
        currency: { type: string }
    Passenger:
      type: object
      required:
        [
          firstName,
          lastName,
          nationality,
          gender,
          dob,
          passportNumber,
          passportExpiry,
        ]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        nationality: { type: string }
        gender: { type: string }
        dob: { type: string, format: date }
        passportNumber: { type: string }
        passportExpiry: { type: string, format: date }
        passportIssueCountry: { type: string }
    FlightBookingStatus:
      type: string
      enum: [pending, confirmed, ticketed, cancelled, completed]
    FlightBooking:
      type: object
      required:
        [
          id,
          userId,
          referenceNumber,
          offer,
          passengers,
          phone,
          email,
          status,
          createdAt,
        ]
      properties:
        id: { type: integer }
        userId: { type: integer }
        referenceNumber: { type: string }
        offer:
          $ref: "#/components/schemas/FlightOffer"
        passengers:
          type: array
          items:
            $ref: "#/components/schemas/Passenger"
        phone: { type: string }
        email: { type: string }
        status:
          $ref: "#/components/schemas/FlightBookingStatus"
        createdAt: { type: string, format: date-time }
    FlightBookingInput:
      type: object
      required: [offer, passengers, phone, email]
      properties:
        offer:
          $ref: "#/components/schemas/FlightOffer"
        passengers:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Passenger"
        phone: { type: string }
        email: { type: string }
    FlightBookingStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/FlightBookingStatus"
    NotificationType:
      type: string
      enum: [visa_application, package_booking, flight_booking, general, promotion]
    Notification:
      type: object
      required: [id, userId, title, message, type, isRead, createdAt]
      properties:
        id: { type: integer }
        userId: { type: ["integer", "null"] }
        title: { type: string }
        message: { type: string }
        type:
          $ref: "#/components/schemas/NotificationType"
        isRead: { type: boolean }
        createdAt: { type: string, format: date-time }
    NotificationInput:
      type: object
      required: [title, message, type]
      properties:
        userId:
          type: ["integer", "null"]
          description: Null broadcasts to all customers
        title: { type: string, minLength: 1 }
        message: { type: string, minLength: 1 }
        type:
          $ref: "#/components/schemas/NotificationType"
...